requirements.txt
streamlit
crewai
openai
requests
beautifulsoup4
pyyaml
